// packages/@higma/editor-controls/src/item-list/types.ts
export type ListItem<TId = string> = { readonly id: TId };

/**
 * Context menu state
 */
export type ItemContextMenuState<TId> = {
  readonly visible: boolean;
  readonly x: number;
  readonly y: number;
  readonly itemId: TId | null;
};